Fix insert for repeated start, end
If multiple vals are inserted into a MultiValueSearchTree with same start, end intervals, insert() will recurse to upsert() rather than insert(), which causes a wholesale replacement of vals rather than appending as it should, resulting in data loss. This causes insert() to recurse to insert() instead, which does the proper append.